Moved LiveObjectManager to internal
