+++ packages/live-share/src/index.ts
@@ -11,7 +11,6 @@ export * from "./LiveEventScope";
 export * from "./LiveEventSource";
 export * from "./LiveEventTarget";
 export * from "./LiveEventTimer";
-export * from "./LiveObjectSynchronizer";
 export * from "./LivePresence";
 export * from "./LivePresenceUser";
 export * from "./LiveShareClient";
@@ -26,3 +25,4 @@ export * from "./schema-utils";
 export * from "./LiveDataObject";
 export * from "./LiveShareRuntime";
 export * from "./AzureLiveShareHost";
+export * from "./LiveObjectSynchronizer";
